Tons of changes
Mainly re-thought how to deal with types and type modifiers such as array levels and pointers.
Added types for every single scalar type, with operators like addition, subtraction, accessing etc.
Added support for swizzle masks on vector types.
Added support for pointers.

+++ code/compiler/v3/ast/structure.h
@@ -8,7 +8,7 @@
 //------------------------------------------------------------------------------
 #include "attribute.h"
 #include "variable.h"
-#include "type.h"
+#include "types/type.h"
 namespace AnyFX
 {
 
@@ -41,8 +41,8 @@ struct Structure : public Type
         Variable::__Resolved::AccessBits accessBits;
 
         static const uint8_t NOT_BOUND = 0xF;
-        uint8_t group;
-        uint8_t binding;
+        uint32_t group;
+        uint32_t binding;
 
         // for alignment
         size_t byteSize;
